# ops-entry/docs/swagger.yaml
basePath: /swagger/index.html
definitions:
  proto.FileResult:
    properties:
      code:
        type: integer
      msg:
        type: string
      request_id:
        type: string
    type: object
host: 0.0.0.0:9090
info:
  contact:
    email: https://gitee.com/openeuler/KubeMate
    name: https://gitee.com/openeuler/KubeMate
    url: https://gitee.com/openeuler/KubeMate
  description: 接受运维管理平台的请求，进行os以云原生的方式进行升级
  license:
    name: Mulan PSL v2
    url: http://license.coscl.org.cn/MulanPSL2
  termsOfService: http://swagger.io/terms/
  title: 统一运维入口
  version: "1.0"
paths:
  /clusterconfig/{cluster_id}:
    delete:
      description: Delete a cluster config file with optional description
      parameters:
      - description: k8s name
        in: path
        name: cluster_id
        required: true
        type: string
      responses:
        "204":
          description: No Content - Indicates successful deletion
      summary: Delete a cluster config file
      tags:
      - 集群配置文件
    get:
      description: Query a clusterconfig file by cluster ID
      parameters:
      - description: k8s cluster ID
        in: path
        name: cluster_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.FileResult'
      summary: Query a clusterconfig file
      tags:
      - 集群配置文件
  /clusterconfig/update:
    put:
      consumes:
      - multipart/form-data
      description: Update a file with optional description
      parameters:
      - description: The cluster config file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: k8s name
        in: formData
        name: cluster_id
        required: true
        type: string
      - description: The JSON string containing labels for the uploaded file
        in: formData
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.FileResult'
      summary: Update a cluster config file
      tags:
      - 集群配置文件
  /clusterconfig/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file with optional description
      parameters:
      - description: The cluster config file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: k8s name
        in: formData
        name: cluster_id
        required: true
        type: string
      - description: The type of the uploaded file (e.g., 'crfile', 'configfile')
        in: formData
        name: type
        required: true
        type: string
      - description: The JSON string containing labels for the uploaded file
        in: formData
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.FileResult'
      summary: Upload a cluster config file
      tags:
      - 集群配置文件
  /kubeconfig/{cluster_id}:
    delete:
      consumes:
      - application/json
      description: Delete a file with optional description
      parameters:
      - description: k8s name
        in: path
        name: cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Indicates successful deletion
      summary: Delete a kubeconfig file
      tags:
      - kubeconfig文件
    get:
      description: Query a kubeconfig file by cluster ID
      parameters:
      - description: k8s cluster ID
        in: path
        name: cluster_id
        required: true
        type: string
      responses:
        "204":
          description: No Content - Indicates successful get
      summary: Query a kubeconfig file
      tags:
      - kubeconfig文件
  /kubeconfig/update:
    put:
      consumes:
      - multipart/form-data
      description: Update a file with optional description
      parameters:
      - description: The kubeconfig file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: k8s name
        in: formData
        name: cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful file update
          schema:
            $ref: '#/definitions/proto.FileResult'
      summary: Update a kubeconfig file
      tags:
      - kubeconfig文件
  /kubeconfig/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file with optional description
      parameters:
      - description: The kubeconfig file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: k8s name
        in: formData
        name: cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful file upload
          schema:
            $ref: '#/definitions/proto.FileResult'
      summary: Upload a kubeconfig file
      tags:
      - kubeconfig文件
swagger: "2.0"